+++ golikov.george/A1/rectangle.hpp
@@ -10,6 +10,6 @@ class Rectangle : public Shape
     double getArea() const override;
     rectangle_t getFrameRect() const override;
     void move(const point_t &pos) override;
-    void move(const float dltX, const float dltY) override;
+    void move(const double dltX, const double dltY) override;
 };
 #endif
